Added option to pad images to 10 sectors/track
Using -p adds a 10th sector to the images written by -o, allowing them
to be written to disks as regular MGT format. This makes reusing the disk
easier without reformatting.

For maximum compatibility the 10th sector in directory entries contain
a space as the first character in the filename. This prevent some programs
from stopping a directory scan early when it thinks the remaining sectors
have never been used.

Padding is disabled if we're writing the special samdos9 image to a data
disk, since it would break the sector count auto-detection mechanism.